apps/rag-script/requirements.txt
psycopg2-binary>=2.9.0
pgvector>=0.2.4
sentence-transformers>=2.2.0
google-generativeai>=0.3.0
numpy>=1.24.0
python-dotenv>=1.0.0
pandas>=1.5.0
scikit-learn>=1.3.0
tiktoken>=0.5.0
tqdm>=4.65.0
pydantic>=2.0.0
typing-extensions>=4.8.0
